Internal

WVZAnalysis.arrow_making — Method

arrow_making(tasks; mapper = map)

Take a collection of tasks, run them via map and mergewith(append!). Returns a dict of vectors representing the datas after filtering.

WVZAnalysis.hist_main — Method

hist_main(tag; mapfun=robust_pmap, no_shape = false, output_dir, kw...)

Running the main looper for a tag (e.g. VH, ttZ) to produce histogram kw... takes anything that AnalysisTask takes.

WVZAnalysis.make_sfsys_wgt! — Function

make_sfsys_wgt!(evt, wgt, nominal_branch, wgt_name, wgt_idx=1)

use wgt_name to find alternative branch names

For every variation, make a new entry in wgt::Dict.

Also multiply nominal for wgt[:NOMINAL]

WVZAnalysis.root_dirs — Method

root_dirs(tag::AbstractString; variation = "sf")

Reutrn a list of dir paths associated with a tag +sf or +shape

WVZAnalysis.set_bdt_model_dir — Method

Set where BDT modesl are (XGBoost)

WVZAnalysis.set_minitree_dir — Method

Set where minitree is.

WVZAnalysis.sumsumWeight — Method

Compute the sum of weights given path to a directory, and cache the result in a text file.

WVZAnalysis.@fill_dict! — Macro

Example:

julia> dd = Dict(:Nlep => [], :lep1_pid=>[]);

julia> Nlep = 10;

julia> lep1_pid =11;

julia> @fill_dict! dd push! Nlep,lep1_pid;

# this expands to
push!(dd[:Nlep], Nlep)
push!(dd[:lep1_pid], lep1_pid)

julia> dd
Dict{Symbol, Vector{Any}} with 2 entries:
  :Nlep     => [10]
  :lep1_pid => [11]

WVZReportExt.print_sigtable — Method

print_sigtable(full_table; io=stdout)

Takes the output of significance_table and pretty print it:

Example

julia> M = significance_table(<path>);

julia> print_sigtable(M)
┌──────────────┬────────────────┬───────────────┬──────────────┬──────────────┬───────────────┐
│              │     SF-inZ     │    SF-noZ     │      DF      │    CR-ZZ     │    CR-ttZ     │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│    Signal    │  10.66 ± 0.07  │  9.31 ± 0.1   │ 10.73 ± 0.14 │ 1.07 ± 0.02  │  2.15 ± 0.04  │
│      ZZ      │ 1219.92 ± 5.38 │ 469.06 ± 2.44 │ 19.78 ± 0.45 │ 555.68 ± 2.4 │ 69.19 ± 0.74  │
│    Zjets     │  -0.02 ± 0.13  │  2.6 ± 2.22   │ 6.47 ± 5.51  │  -0.0 ± 0.0  │  0.63 ± 0.39  │
│    Zgamma    │   0.0 ± 0.0    │   0.0 ± 0.0   │  0.3 ± 0.29  │  0.0 ± 0.0   │   0.0 ± 0.0   │
│    ttbar     │   0.0 ± 0.0    │  0.63 ± 0.18  │  0.28 ± 0.1  │  0.0 ± 0.0   │  0.43 ± 0.13  │
│      WZ      │   0.36 ± 0.1   │  1.79 ± 0.23  │ 2.24 ± 0.29  │  0.0 ± 0.0   │  0.29 ± 0.06  │
│      tZ      │  0.01 ± 0.01   │  0.07 ± 0.03  │ 0.06 ± 0.02  │  0.0 ± 0.0   │  0.18 ± 0.04  │
│     ttZ      │  1.23 ± 0.08   │  4.71 ± 0.16  │ 5.74 ± 0.18  │ 0.01 ± 0.01  │ 73.62 ± 0.61  │
│     tWZ      │  0.56 ± 0.11   │  2.16 ± 0.23  │  2.5 ± 0.24  │ 0.01 ± 0.01  │ 14.71 ± 0.56  │
│     VBS      │  10.23 ± 0.09  │  6.4 ± 0.08   │ 0.18 ± 0.01  │ 1.48 ± 0.03  │  0.83 ± 0.02  │
│      VH      │  1.29 ± 0.71   │  5.76 ± 1.4   │ 5.77 ± 1.29  │  0.0 ± 0.0   │   0.0 ± 0.0   │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│    Others    │  0.06 ± 0.01   │  0.4 ± 0.13   │ 0.56 ± 0.08  │  0.0 ± 0.0   │  5.02 ± 0.09  │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│   Bkg Tot.   │ 1233.64 ± 5.43 │ 493.58 ± 3.61 │ 43.89 ± 5.7  │ 557.19 ± 2.4 │ 164.91 ± 1.19 │
│ Significance │   0.3 ± 0.0    │  0.42 ± 0.0   │  1.56 ± 0.1  │  0.05 ± 0.0  │  0.17 ± 0.0   │
└──────────────┴────────────────┴───────────────┴──────────────┴──────────────┴───────────────┘

WVZReportExt.rebinscan — Method

rebinscan(S, B; atleast=1, from=:right, by = (s,b) -> s/sqrt(b))

Take two Hist1D, try to find the best bin-edges for rebinning, given these two conditions:

a metric to maximize, s and b are signal counts and background counts in each new bin

atleast specifcy the minimal number of s+b in each newly formed bin

from keyword argument can be used to specifcy the direction of the scan, defautls to :right assuming the high-er end of the histogram is signal-like and one of the histograms is monotonic.

The function returns the values of new bin-edges including both ends (of course, the ends are identical to before).

The metric can also be replaced by the more accurate sqrt(2*((s + b) * log(1 + s/b) - s )).

WVZReportExt.significance_table — Method

significance_table()
significance_table(significance_matrix(); )

Examples

julia> M = significance_table()
14×6 Matrix{Any}:
 "Signal"         10.66±0.067   …    1.072±0.017     2.151±0.037
 "ZZ"            1219.9±5.4          555.7±2.4       69.19±0.74
 "Zjets"         -0.019±0.13       -0.0004±0.0004     0.63±0.39
 "Zgamma"           0.0±0.0            0.0±0.0         0.0±0.0
 "ttbar"            0.0±0.0            0.0±0.0        0.43±0.13
 "WZ"             0.358±0.096   …   0.0044±0.0032    0.295±0.061
 "tZ"             0.012±0.012          0.0±0.0       0.179±0.043
 "ttZ"            1.231±0.08        0.0111±0.0074    73.62±0.61
 "tWZ"             0.56±0.11        0.0095±0.0095    14.71±0.56
 "VBS"           10.231±0.085        1.478±0.033     0.833±0.024
 "VH"              1.29±0.71    …      0.0±0.0         0.0±0.0
 "Others"        0.0568±0.0088      0.0021±0.0017     5.02±0.088
 "Bkg Tot."      1233.6±5.4          557.2±2.4       164.9±1.2
 "Significance"  0.3031±0.002       0.0454±0.00072  0.1672±0.0029