The Big Picture

like any Julia package, the source files are under /src

meta data files such as tag -> DSID mapping are under /config

the main looper is called main_looper() and all the real actions are in there

the atomic unit of runnable task is called AnalysisTask

A typical workflow

Say we're trying to produce BDT score histogram for process ZZ:

call prep_tasks("ZZ") (this returns a vector of AnalysisTasks)

run main_looper over each of the tasks, you can use threading, or distributed parallelism.

each main_looper would would return a dictionary of histograms

merge all the result (reduce(mergewith(+), results_from_3))

profit

Public-ish APIs

These are the types and functions all a user would need, in order to do pretty much anything.

The packages that are good to be familiar with:

UnROOT.jl

FHist.jl

These functions/types and the direct callees are more documented than the rest of the library:

Analysis and data dumping

WVZAnalysis.AnalysisTask — Type

struct AnalysisTask
    path::String
    sumWeight::Float64
    isdata::Bool = false
    BDT_hist::Bool = true
    arrow_making::Bool = false
    sfsys::Bool = false
    shape_variation::String = "NOMINAL"
    controlregion::Symbol = :none
end

Fully define a task to be run on an executor, by calling main_looper(task::AnalysisTask); see also main_looper.

You most likely don't need to construct it manually, see prep_tasks:

Example

julia> prep_tasks("Signal") |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
BDT_hist=true
arrow_making=false
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

julia> prep_tasks("Signal"; arrow_making=true) |> first
ERROR: can't do produce arrow and NN histograms at the same time
Stacktrace:
...
..

julia> prep_tasks("Signal"; arrow_making=true, BDT_hist=false) |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
BDT_hist=false
arrow_making=true
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

WVZAnalysis.prep_tasks — Function

prep_tasks(tag; shape_variation="NOMINAL", scouting=false, kw...)

For construction a collection of AnalysisTasks given tag and job options, check out [@AnalysisTask].

All posible tags can be found in config/file_list.json, there's also a convinient variable WVZAnalysis.ALL_TAGS that keeps track of all processes in an exclusive mannar:

Example

julia> WVZAnalysis.ALL_TAGS
("Signal", "ZZ", "Zjets", "Zgamma", "ttbar", "WZ", "tZ", "ttZ", "tWZ", "VBS", "VH", "Others")

julia> all_nominal_tasks = mapreduce(prep_tasks, vcat, WVZAnalysis.ALL_TAGS);

julia> length(all_nominal_tasks)
768

WVZAnalysis.main_looper — Method

main_looper(task::AnalysisTask)

The main entry point for running the main looper for a given task, it's done in steps:

destruct all the options from the task (see AnalysisTask):

    (; path, sumWeight, arrow_making, BDT_hist, isdata, 
     shape_variation, controlregion, sfsys) = task

determine what output dict to prepare.

call the lower level main_looper (which has all arguments explicitly laied out)

This function also serves a function barrier for performance reason, because we have so many different behaviors in the same loop function.

Missing docstring.

Missing docstring for arrow_making. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hist_root. Check Documenter's build log for details.

Result reporting

Missing docstring.

Missing docstring for significance_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for print_sigtable. Check Documenter's build log for details.