The Big Picture

like any Julia package, the source files are under /src

meta data files such as tag -> DSID mapping are under /config

the main looper is called main_looper() and all the real actions are in there

the atomic unit of runnable task is called AnalysisTask

A typical workflow

Say we're trying to produce BDT score histogram for process ZZ:

call prep_tasks("ZZ") (this returns a vector of AnalysisTasks)

run main_looper over each of the tasks, you can use threading, or distributed parallelism.

each main_looper would would return a dictionary of histograms

merge all the result (reduce(mergewith(+), results_from_3))

profit

Public-ish APIs

These are the types and functions all a user would need, in order to do pretty much anything.

The packages that are good to be familiar with:

UnROOT.jl

FHist.jl

These functions/types and the direct callees are more documented than the rest of the library:

Analysis and data dumping

WVZAnalysis.AnalysisTask — Type

struct AnalysisTask
    path::String
    sumWeight::Float64
    isdata::Bool = false
    NN_hist::Bool = true
    arrow_making::Bool = false
    sfsys::Bool = false
    shape_variation::String = "NOMINAL"
    controlregion::Symbol = :none
end

Fully define a task to be run on an executor, by calling main_looper(task::AnalysisTask); see also main_looper.

You most likely don't need to construct it manually, see prep_tasks:

Example

julia> prep_tasks("Signal") |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
NN_hist=true
arrow_making=false
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

julia> prep_tasks("Signal"; arrow_making=true) |> first
ERROR: can't do produce arrow and NN histograms at the same time
Stacktrace:
...
..

julia> prep_tasks("Signal"; arrow_making=true, NN_hist=false) |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
NN_hist=false
arrow_making=true
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

WVZAnalysis.prep_tasks — Function

prep_tasks(tag; shape_variation="NOMINAL", scouting=false, kw...)

For construction a collection of AnalysisTasks given tag and job options, check out [@AnalysisTask].

All posible tags can be found in config/file_list.json, there's also a convinient variable WVZAnalysis.ALL_TAGS that keeps track of all processes in an exclusive mannar:

Example

julia> WVZAnalysis.ALL_TAGS
("Signal", "ZZ", "Zjets", "Zgamma", "ttbar", "WZ", "tZ", "ttZ", "tWZ", "VBS", "VH", "Others")

julia> all_nominal_tasks = mapreduce(prep_tasks, vcat, WVZAnalysis.ALL_TAGS);

julia> length(all_nominal_tasks)
768

WVZAnalysis.main_looper — Method

main_looper(task::AnalysisTask)

The main entry point for running the main looper for a given task, it's done in steps:

destruct all the options from the task (see AnalysisTask):

    (; path, sumWeight, arrow_making, NN_hist, isdata, 
     shape_variation, controlregion, sfsys) = task

determine what output dict to prepare.

call the lower level main_looper (which has all arguments explicitly laied out)

This function also serves a function barrier for performance reason, because we have so many different behaviors in the same loop function.

WVZAnalysis.arrow_making — Function

arrow_making(tasks)

Take a collection of tasks, run them via map and mergewith(append!). Returns a dict of vectors representing the datas after filtering.

WVZAnalysis.hist_root — Function

hist_root(tag; output_dir, kw...)

return a dictionary of all systematics histograms for a given tag, the output_dir is used to store Julia Serializatio files as a backup measure, and for later conversion to .root histograms via PythonCall.jl + uproot.

WVZAnalysis.hist_root_pmap — Function

hist_root_pmap(tag; output_dir, kw...)

Similar to the one without pmap, except uses pmap for everything. Checkout ClusterManager.jl and be sure to have a handful of workers before running this.

Result reporting

WVZAnalysis.significance_table — Function

significance_table(; recreate=false)
significance_table(significance_matrix(); recreate=false)

Examples

julia> M = significance_table()
14×6 Matrix{Any}:
 "Signal"         10.66±0.067   …    1.072±0.017     2.151±0.037
 "ZZ"            1219.9±5.4          555.7±2.4       69.19±0.74
 "Zjets"         -0.019±0.13       -0.0004±0.0004     0.63±0.39
 "Zgamma"           0.0±0.0            0.0±0.0         0.0±0.0
 "ttbar"            0.0±0.0            0.0±0.0        0.43±0.13
 "WZ"             0.358±0.096   …   0.0044±0.0032    0.295±0.061
 "tZ"             0.012±0.012          0.0±0.0       0.179±0.043
 "ttZ"            1.231±0.08        0.0111±0.0074    73.62±0.61
 "tWZ"             0.56±0.11        0.0095±0.0095    14.71±0.56
 "VBS"           10.231±0.085        1.478±0.033     0.833±0.024
 "VH"              1.29±0.71    …      0.0±0.0         0.0±0.0
 "Others"        0.0568±0.0088      0.0021±0.0017     5.02±0.088
 "Bkg Tot."      1233.6±5.4          557.2±2.4       164.9±1.2
 "Significance"  0.3031±0.002       0.0454±0.00072  0.1672±0.0029

WVZAnalysis.print_sigtable — Function

print_sigtable(full_table)

Takes the output of significance_table and pretty print it:

Example

julia> M = significance_table(; recreate=true);

julia> print_sigtable(M)
┌──────────────┬────────────────┬───────────────┬──────────────┬──────────────┬───────────────┐
│              │     SF-inZ     │    SF-noZ     │      DF      │    CR-ZZ     │    CR-ttZ     │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│    Signal    │  10.66 ± 0.07  │  9.31 ± 0.1   │ 10.73 ± 0.14 │ 1.07 ± 0.02  │  2.15 ± 0.04  │
│      ZZ      │ 1219.92 ± 5.38 │ 469.06 ± 2.44 │ 19.78 ± 0.45 │ 555.68 ± 2.4 │ 69.19 ± 0.74  │
│    Zjets     │  -0.02 ± 0.13  │  2.6 ± 2.22   │ 6.47 ± 5.51  │  -0.0 ± 0.0  │  0.63 ± 0.39  │
│    Zgamma    │   0.0 ± 0.0    │   0.0 ± 0.0   │  0.3 ± 0.29  │  0.0 ± 0.0   │   0.0 ± 0.0   │
│    ttbar     │   0.0 ± 0.0    │  0.63 ± 0.18  │  0.28 ± 0.1  │  0.0 ± 0.0   │  0.43 ± 0.13  │
│      WZ      │   0.36 ± 0.1   │  1.79 ± 0.23  │ 2.24 ± 0.29  │  0.0 ± 0.0   │  0.29 ± 0.06  │
│      tZ      │  0.01 ± 0.01   │  0.07 ± 0.03  │ 0.06 ± 0.02  │  0.0 ± 0.0   │  0.18 ± 0.04  │
│     ttZ      │  1.23 ± 0.08   │  4.71 ± 0.16  │ 5.74 ± 0.18  │ 0.01 ± 0.01  │ 73.62 ± 0.61  │
│     tWZ      │  0.56 ± 0.11   │  2.16 ± 0.23  │  2.5 ± 0.24  │ 0.01 ± 0.01  │ 14.71 ± 0.56  │
│     VBS      │  10.23 ± 0.09  │  6.4 ± 0.08   │ 0.18 ± 0.01  │ 1.48 ± 0.03  │  0.83 ± 0.02  │
│      VH      │  1.29 ± 0.71   │  5.76 ± 1.4   │ 5.77 ± 1.29  │  0.0 ± 0.0   │   0.0 ± 0.0   │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│    Others    │  0.06 ± 0.01   │  0.4 ± 0.13   │ 0.56 ± 0.08  │  0.0 ± 0.0   │  5.02 ± 0.09  │
├──────────────┼────────────────┼───────────────┼──────────────┼──────────────┼───────────────┤
│   Bkg Tot.   │ 1233.64 ± 5.43 │ 493.58 ± 3.61 │ 43.89 ± 5.7  │ 557.19 ± 2.4 │ 164.91 ± 1.19 │
│ Significance │   0.3 ± 0.0    │  0.42 ± 0.0   │  1.56 ± 0.1  │  0.05 ± 0.0  │  0.17 ± 0.0   │
└──────────────┴────────────────┴───────────────┴──────────────┴──────────────┴───────────────┘