The Big Picture

like any Julia package, the source files are under /src

meta data files such as tag -> DSID mapping are under /config

the main looper is called main_looper() and all the real actions are in there

the atomic unit of runnable task is called AnalysisTask

A typical workflow

Say we're trying to produce BDT score histogram for process ZZ:

call prep_tasks("ZZ") (this returns a vector of AnalysisTasks)

run main_looper over each of the tasks, you can use threading, or distributed parallelism.

each main_looper would would return a dictionary of histograms

merge all the result (reduce(mergewith(+), results_from_3))

profit

Public-ish APIs

These are the types and functions all a user would need, in order to do pretty much anything.

The packages that are good to be familiar with:

UnROOT.jl

FHist.jl

These functions/types and the direct callees are more documented than the rest of the library:

Analysis and data dumping

WVZAnalysis.AnalysisTask — Type

struct AnalysisTask
    path::String
    sumWeight::Float64
    isdata::Bool = false
    NN_hist::Bool = true
    arrow_making::Bool = false
    sfsys::Bool = false
    shape_variation::String = "NOMINAL"
    controlregion::Symbol = :none
end

Fully define a task to be run on an executor, by calling main_looper(task::AnalysisTask); see also main_looper.

You most likely don't need to construct it manually, see prep_tasks:

Example

julia> prep_tasks("Signal") |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
NN_hist=true
arrow_making=false
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

julia> prep_tasks("Signal"; arrow_making=true) |> first
ERROR: can't do produce arrow and NN histograms at the same time
Stacktrace:
...
..

julia> prep_tasks("Signal"; arrow_making=true, NN_hist=false) |> first
path="/data/jiling/WVZ/v2.3/user.jiling.WVZ_v2.3sf.363507.e6379_s3126_r10201_p4434_ANALYSIS.root/user.jiling.29896106._000001.ANALYSIS.root"
sumWeight=13812.79638671875
isdata=false
NN_hist=false
arrow_making=true
sfsys=false
shape_variation="NOMINAL"
controlregion=:none

WVZAnalysis.prep_tasks — Function

prep_tasks(tag; shape_variation="NOMINAL", scouting=false, kw...)

For construction a collection of AnalysisTasks given tag and job options, check out [@AnalysisTask].

All posible tags can be found in config/file_list.json, there's also a convinient variable WVZAnalysis.ALL_TAGS that keeps track of all processes in an exclusive mannar:

julia> WVZAnalysis.ALL_TAGS
("Signal", "ZZ", "Zjets", "Zgamma", "ttbar", "WZ", "tZ", "ttZ", "tWZ", "VBS", "VH", "Others")

Example

julia> all_nominal_tasks = mapreduce(prep_tasks, vcat, WVZAnalysis.ALL_TAGS);

julia> length(all_nominal_tasks)
768

WVZAnalysis.main_looper — Method

main_looper(task::AnalysisTask)

The main entry point for running the main looper for a given task, it's done in steps:

destruct all the options from the task (see AnalysisTask):

    (; path, sumWeight, arrow_making, NN_hist, isdata, 
     shape_variation, controlregion, sfsys) = task

determine what output dict to prepare.

call the lower level main_looper (which has all arguments explicitly laied out)

This function also serves a function barrier for performance reason, because we have so many different behaviors in the same loop function.

WVZAnalysis.arrow_making — Function

arrow_making(tasks)

Take a collection of tasks, run them via ThreadsX.map and mergewith(append!). Returns a dict of vectors representing the datas after filtering.

WVZAnalysis.hist_root — Function

hist_root(tag; output_dir, kw...)

return a dictionary of all systematics histograms for a given tag, the output_dir is used to store Julia Serializatio files as a backup measure, and for later conversion to .root histograms via PythonCall.jl + uproot.

WVZAnalysis.hist_root_pmap — Function

hist_root_pmap(tag; output_dir, kw...)

Similar to the one without pmap, except uses pmap for everything. Checkout ClusterManager.jl and be sure to have a handful of workers before running this.

Result reporting

WVZAnalysis.significance_table — Function

significance_table(; recreate=false)
significance_table(significance_matrix(); recreate=false)

Examples

julia> M = significance_table()
15×4 Matrix{Any}:
 "Signal"         11.509±0.068     9.18±0.1     10.14±0.14
 "ZZ"             1709.1±5.5      389.2±2.3      19.1±0.44
 "Zjets"           -0.02±0.13       1.9±1.1       5.9±5.5
 "Zgamma"            0.0±0.0        0.0±0.0       0.3±0.29
 "ttbar"             0.0±0.0       0.63±0.18     0.28±0.1
 "WZ"              0.362±0.096     1.72±0.22     2.13±0.28
 "tZ"                0.0±0.0      0.059±0.025    0.06±0.025
 "ttZ"             1.227±0.08      4.66±0.16     5.68±0.17
 "tWZ"              0.58±0.11      2.15±0.23     2.51±0.24
 "VBS"            11.682±0.092    5.239±0.073   0.182±0.011
 "VH"               1.29±0.71       5.6±1.4       5.4±1.2
 "Others"         0.0584±0.0089    0.38±0.13    0.549±0.082
 "Bkg Tot."       1724.3±5.6      411.5±2.9      42.1±5.7
 "Significance"   0.2768±0.0017  0.4508±0.0053  1.505±0.096
 "Combined Sig."     NaN±0.0      1.595±0.091     NaN±0.0

WVZAnalysis.print_sigtable — Function

print_sigtable(full_table)

Takes the output of significance_table and pretty print it:

Example

julia> M = significance_table(; recreate=true);

julia> print_sigtable(M)

┌───────────────┬────────────────┬───────────────┬──────────────┐
│               │     SF-inZ     │    SF-noZ     │      DF      │
├───────────────┼────────────────┼───────────────┼──────────────┤
│    Signal     │  11.73 ± 0.07  │  9.31 ± 0.1   │ 10.73 ± 0.14 │
│      ZZ       │ 1775.6 ± 5.89  │ 469.06 ± 2.44 │ 19.78 ± 0.45 │
│     Zjets     │  -0.02 ± 0.13  │  2.6 ± 2.22   │ 6.47 ± 5.51  │
│    Zgamma     │   0.0 ± 0.0    │   0.0 ± 0.0   │  0.3 ± 0.29  │
│     ttbar     │   0.0 ± 0.0    │  0.63 ± 0.18  │  0.28 ± 0.1  │
│      WZ       │   0.36 ± 0.1   │  1.79 ± 0.23  │ 2.24 ± 0.29  │
│      tZ       │  0.01 ± 0.01   │  0.07 ± 0.03  │ 0.06 ± 0.02  │
│      ttZ      │  1.24 ± 0.08   │  4.71 ± 0.16  │ 5.74 ± 0.18  │
│      tWZ      │  0.57 ± 0.11   │  2.16 ± 0.23  │  2.5 ± 0.24  │
│      VBS      │  11.71 ± 0.09  │  6.4 ± 0.08   │ 0.18 ± 0.01  │
│      VH       │  1.29 ± 0.71   │  5.76 ± 1.4   │ 5.77 ± 1.29  │
│    Others     │  0.06 ± 0.01   │  0.4 ± 0.13   │ 0.56 ± 0.08  │
├───────────────┼────────────────┼───────────────┼──────────────┤
│   Bkg Tot.    │ 1790.82 ± 5.94 │ 493.58 ± 3.61 │ 43.89 ± 5.7  │
├───────────────┼────────────────┼───────────────┼──────────────┤
│ Significance  │   0.28 ± 0.0   │  0.42 ± 0.0   │  1.56 ± 0.1  │
│ Combined Sig. │   NaN ± 0.0    │  1.64 ± 0.09  │  NaN ± 0.0   │
└───────────────┴────────────────┴───────────────┴──────────────┘